se agrega auth a todas las vistas y se agrega la vista de filtro para nombre de tareas

+++ todoApp/views.py
@@ -27,6 +27,17 @@ def apiOverview(request):
     return Response (api_urls)
 
 
+""" se deja para poder tener acceso a todas la tareas sin tener
+#          encuenta de q usuario son por las dudas """
+
+# @api_view(['GET'])
+# def listaTareas(request):
+#     tareas = Tarea.objects.all()
+#     serializer = TareaSerializer(tareas, many=True)
+#     print(serializer)
+#     return Response(serializer.data)
+
+
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
 def listaTareas(request):
@@ -35,32 +46,60 @@ def listaTareas(request):
     return Response(serializer.data)
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def detallesTarea(request, pk):
-    tareas = Tarea.objects.get(id=pk)
-    serializer = TareaSerializer(tareas, many=False)
-    return Response(serializer.data)
+    tarea = Tarea.objects.get(id=pk)
+    if (tarea.usuario == request.user):
+        serializer = TareaSerializer(tarea, many=False)
+        return Response(serializer.data)
+    else:
+        return Response('Tarea no pertenece al usuario logueado')
+
+
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def crearTarea(request):
-    serializer = TareaSerializer(data=request.data)
+    datos = request.data
+    datos['usuario'] = request.user.id
+    serializer = TareaSerializer(data=datos)
+    print(request.data)
     if serializer.is_valid():
         serializer.save()
-    return Response(serializer.data)
+        return Response(serializer.data)
+    else:
+        print(serializer.errors)
+
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def modificarTarea(request, pk):
     tarea = Tarea.objects.get(id=pk)
-    serializer = TareaSerializer(instance=tarea, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-    return Response(serializer.data)
+    if (tarea.usuario == request.user):
+        serializer = TareaSerializer(instance=tarea, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+        return Response(serializer.data)
+    else:
+        return Response('Tarea no pertenece al usuario logueado')
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def borrarTarea(request, pk):
     tarea = Tarea.objects.get(id=pk)
-    tarea.delete()
-    return Response('Tarea eliminada!')
+    if (tarea.usuario == request.user):
+        tarea.delete()
+        return Response('Tarea eliminada!')
+    else:
+        return Response('Tarea no pertenece al usuario logueado')
 
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def filtrarContenido (request, fil):
+    tareas = Tarea.objects.filter(tarea__icontains=fil, usuario=request.user)
+    serializer = TareaSerializer(tareas, many=True)
+    return Response(serializer.data)
 
 #token views modificadas
